разбил на файлы, добавил выбор башен
