+++ index.html
@@ -12,6 +12,6 @@
 <body>
   <canvas id="canvas1"></canvas>
 
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
